translate_in_many_style/requirements.txt
streamlit
streamlit-chat
websocket-client